requirements.txt
numpy>=1.24
scipy>=1.10
scikit-image>=0.22
matplotlib>=3.7
tqdm>=4.66
pyyaml>=6.0
torch>=2.2
torchvision>=0.17
pytest>=7.4
